Move select to data_transfer
